ai/requirements.txt
# Core Dependencies
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI & Machine Learning
openai>=1.3.0
groq>=0.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Natural Language Processing
nltk>=3.8.0
spacy>=3.7.0

# Video Generation & Multimedia
moviepy>=1.0.3
pyttsx3>=2.90
matplotlib>=3.7.0

# Audio/Video Backends (for moviepy)
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
Pillow>=10.0.0

# Additional Utilities
uuid>=1.30
asyncio>=3.4.3
typing-extensions>=4.8.0